pyproject.toml
[tool.poetry]
name = "klm-techcase"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.124.4"}
sqlmodel = "^0.0.27"
alembic = "^1.17.2"
psycopg2-binary = "^2.9.11"
sqlalchemy-utils = "^0.42.0"
httpx = "^0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-docker = "^3.2.5"
pytest-mock = "^3.15.1"
ruff = "^0.14.9"

[tool.poetry.scripts]
klm-docs = "klm_techcase.main:generate_docs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
include = ["*.py"]
line-length = 88
indent-width = 4
fix = true
target-version = "py310"
